Application_Code_Files/requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.1
airportsdata==20241001
alembic==1.14.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.7.0
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.13.3
blake3==1.0.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
compressed-tensors==0.8.1
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.41.0
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.9
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
einops==0.8.0
executing==2.1.0
faiss-cpu==1.9.0.post1
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
fonttools==4.55.7
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
groq==0.13.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflection==0.5.1
interegular==0.3.3
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.8
langchain==0.3.11
langchain-community==0.3.11
langchain-core==0.3.24
langchain-experimental==0.3.3
langchain-google-genai==2.0.7
langchain-groq==0.2.1
langchain-openai==0.2.12
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.2.3
lark==1.2.2
latitude-sdk==1.0.0
latitude-telemetry==1.0.0
libclang==18.1.1
lm-format-enforcer==0.10.9
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.1
ml-dtypes==0.4.1
mlflow==2.20.0
mlflow-skinny==2.20.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.15.2
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-ml-py==12.560.30
openai==1.57.3
opencv-python-headless==4.10.0.84
openinference-instrumentation==0.1.21
openinference-instrumentation-dspy==0.1.14
openinference-instrumentation-litellm==0.1.6
openinference-semantic-conventions==0.1.12
opentelemetry-api==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-alephalpha==0.37.1
opentelemetry-instrumentation-anthropic==0.37.1
opentelemetry-instrumentation-bedrock==0.37.1
opentelemetry-instrumentation-cohere==0.37.1
opentelemetry-instrumentation-google-generativeai==0.37.1
opentelemetry-instrumentation-groq==0.37.1
opentelemetry-instrumentation-haystack==0.37.1
opentelemetry-instrumentation-langchain==0.37.1
opentelemetry-instrumentation-llamaindex==0.37.1
opentelemetry-instrumentation-mistralai==0.37.1
opentelemetry-instrumentation-ollama==0.37.1
opentelemetry-instrumentation-openai==0.37.1
opentelemetry-instrumentation-replicate==0.37.1
opentelemetry-instrumentation-sagemaker==0.37.1
opentelemetry-instrumentation-threading==0.50b0
opentelemetry-instrumentation-together==0.37.1
opentelemetry-instrumentation-transformers==0.37.1
opentelemetry-instrumentation-vertexai==0.37.1
opentelemetry-instrumentation-watsonx==0.37.1
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-semantic-conventions-ai==0.4.2
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.12
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
promptl-ai==1.0.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.1
rsa==4.9
safetensors==0.4.5
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.1
sentencepiece==0.2.0
setuptools==75.6.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.41.3
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
vllm==0.6.5
waitress==3.0.2
wasmtime==29.0.0
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
